IDAM-253 Add Eslint and prettier setup, file restructure
