+++ NICE.Identity.Management/ClientApp/.prettierrc
@@ -1,0 +1,18 @@
+{
+  "semi": true,
+  "singleQuote": false,
+  "trailingComma": "all",
+  "useTabs": true,
+  "printWidth": 80,
+  "endOfLine": "lf",
+  "overrides": [
+    {
+      "files": ["**/*.scss"],
+      "options": {
+        "useTabs": false,
+        "tabWidth": 2,
+        "singleQuote": true
+      }
+    }
+  ]
+}
