+++ NICE.Identity.Management/ClientApp/src/components/App/App.scss
