+++ NICE.Identity.Management/ClientApp/src/components/App/__tests__/App.test.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "../App";
+
+it("renders without crashing", () => {
+	const div = document.createElement("div");
+	ReactDOM.render(<App />, div);
+	ReactDOM.unmountComponentAtNode(div);
+});
