+++ NICE.Identity.Management/ClientApp/src/index.scss
